Refactored PEMS agglomeration to reduce memory usage
Because the agglomerator read in the full set of PEMS data requested
before agglomerating, memory usage would increase with the length
or density of the dataset desired. For the 5-minute PEMS data,
this can be ~300 MB per district per day, which is untenable. This
commit completely refactors the internal logic of the agglomerator
so that only one file is held in memory at a time and each file's
data is written to the netCDF file before the next is loaded.