Memory optimization changed to allow for counties in multiple districts
My previous assumption that all stations for one county would be in a
single district appears to have been wrong. So now the variables are
totaled in a running sum for all files and only written to the netCDF
file at the end. This is less flexible regarding how the results are
calculated, but allows contributions to a single data point from
multiple files.